feat: add navbar to router outlet
