+++ src/components/Layout.tsx
@@ -1,26 +1,12 @@
-import { Link, Outlet } from "react-router-dom";
+import { Outlet } from "react-router-dom";
+import Navbar from "./ui/navbar";
 
 const Layout = () => {
   return (
     <>
-      <div className="max-w-[1200px] m-auto p-2 md:py-4 flex flex-col md:flex-row gap-4">
-        <nav className="w-[200px] pt-6">
-          <ul>
-            <li>
-              <Link to="/home" viewTransition>Home</Link>
-            </li>
-
-            <li>
-              <Link to="/profile" viewTransition>Profile</Link>
-            </li>
-            
-            <li>
-              <Link to="/projects" viewTransition>Projects</Link>
-            </li>
-          </ul>
-        </nav>
-
-        <div className="w-full pt-4">
+      <Navbar />
+      <div className="sm:max-w-[800px] xl:max-w-[1000px] m-auto p-2 md:py-4 flex flex-col md:flex-row gap-4">
+        <div className="w-full px-2 pt-4 mt-[80px]">
           <Outlet />
         </div>
       </div>
